To run the php code, the website should be launched from the lamp server.
In the file "php_connect.php", replace the database_username, database_password,  and database_name strings to be able to connect to the database.

Use the "lamp.cse.fau.edu/~username/php/make_tables.php" address to create the tables once the "php_connect.php" file is corrected.

As of now, a user can access the home page, dashboard page, about page, and tutorial pages. Content for the tutorial and about pages are not typed yet.

The website has an account system which enabled the user to signup and login into the website using email account activation and a database.

For the account activation, in the file "parser.php" on line 91, the FROM email variable should be changed to a working email so the activation email does not get sent to the junk folder.

The design is open for any kinds of change.